requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
mutagen==1.47.0
pyyaml==6.0.1
python-multipart==0.0.6